+++ src/views/admin/Event/EventsListView.vue
