+++ src/views/auth/SignInView.vue
@@ -67,8 +67,8 @@
 </template>
 
 <script>
-import navbar from "@/components/navbar.vue";
-import Footer from "@/components/footer.vue";
+import navbar from "@/components/Navbar.vue";
+import Footer from "@/components/Footer.vue";
 import router from "@/router";
 import { useUserStore } from "@/store/user";
