+++ src/views/auth/SignUpView.vue
@@ -150,8 +150,8 @@
 </template>
 
 <script>
-import navbar from "@/components/navbar.vue";
-import Footer from "@/components/footer.vue";
+import navbar from "@/components/Navbar.vue";
+import Footer from "@/components/Footer.vue";
 import router from "@/router";
 import { useUserStore } from "@/store/user";
