+++ src/views/event/events.vue
@@ -188,7 +188,7 @@
 </template>
 
 <script setup>
-import navbar from "@/components/navbar.vue";
+import navbar from "@/components/Navbar.vue";
 import FilterPopup from "@/components/FilterPopup.vue";
 import { ref, computed, onMounted } from "vue";
 import { useUserStore } from "@/store/user";
