--- src/views/registerEvent.vue
@@ -1,26 +1,0 @@
-<template>
-  <navbar />
-  <div class="container mt-3 mb-5"></div>
-</template>
-
-<script setup>
-import navbar from "@/components/navbar.vue";
-import Footer from "@/components/footer.vue";
-import { ref, computed, watch, reactive } from "vue";
-</script>
-
-<style scoped>
-p {
-  font-family: "Font";
-}
-
-h4 {
-  font-family: "Font";
-}
-li {
-  font-family: "Font";
-}
-h1 {
-  font-family: "Font";
-}
-</style>
